requirements.txt
psycopg2-binary
pandas
openpyxl
requests
beautifulsoup4
lxml
PySide6